+++ test/host/device_config_test/main/test_device_config.c
@@ -58,6 +58,8 @@ TEST_CASE("defaults match SPEC.md section 9", "[device_config]")
         TEST_ASSERT_EQUAL(1000, c.relays[i].pulse_ms);
         TEST_ASSERT_EQUAL(FAILSAFE_DEFAULT_SAFE, c.relays[i].failsafe);
         TEST_ASSERT_EQUAL(0, c.relays[i].override_di);
+        TEST_ASSERT_EQUAL(INPUT_LINK_FOLLOW, c.relays[i].link);
+        TEST_ASSERT_EQUAL(0, c.relays[i].max_on_s);
         TEST_ASSERT_FALSE(c.inputs[i].invert);
     }
     store_down();
@@ -73,6 +75,8 @@ TEST_CASE("stored values are read into the right channel", "[device_config]")
     TEST_ESP_OK(espos_config_set_str("swbank", "relay5_failsafe", "hold"));
     TEST_ESP_OK(espos_config_set_i32("swbank", "relay5_override", 2));
     TEST_ESP_OK(espos_config_set_bool("swbank", "input2_invert", true));
+    TEST_ESP_OK(espos_config_set_str("swbank", "relay5_link", "toggle"));
+    TEST_ESP_OK(espos_config_set_i32("swbank", "relay6_max_on_s", 1800));
 
     device_config_t c;
     TEST_ESP_OK(device_config_load(&c));
@@ -84,6 +88,9 @@ TEST_CASE("stored values are read into the right channel", "[device_config]")
     TEST_ASSERT_EQUAL(2, c.relays[4].override_di);
     TEST_ASSERT_TRUE(c.inputs[1].invert);
     TEST_ASSERT_FALSE(c.inputs[0].invert);
+    TEST_ASSERT_EQUAL(INPUT_LINK_TOGGLE, c.relays[4].link);
+    TEST_ASSERT_EQUAL(INPUT_LINK_FOLLOW, c.relays[5].link);
+    TEST_ASSERT_EQUAL(1800, c.relays[5].max_on_s);
     store_down();
 }
 
@@ -93,6 +100,8 @@ TEST_CASE("the store rejects out-of-range values", "[device_config]")
     TEST_ASSERT_NOT_EQUAL(ESP_OK, espos_config_set_i32("swbank", "bank_id", 253));
     TEST_ASSERT_NOT_EQUAL(ESP_OK, espos_config_set_i32("swbank", "relay1_override", 9));
     TEST_ASSERT_NOT_EQUAL(ESP_OK, espos_config_set_str("swbank", "relay1_mode", "toggle"));
+    TEST_ASSERT_NOT_EQUAL(ESP_OK, espos_config_set_str("swbank", "relay1_link", "latch"));
+    TEST_ASSERT_NOT_EQUAL(ESP_OK, espos_config_set_i32("swbank", "relay1_max_on_s", 86401));
     store_down();
 }